refactor: standardize DTO naming and package structure
This commit refactors the data layer to provide a more consistent naming convention for Data Transfer Objects (DTOs), standardizing on the `Dto` suffix and moving them into dedicated `dto` packages. It also eliminates redundant data classes across different features.

- **Standardized Naming**:
    - Renamed `UserItem` to `UserDto`.
    - Renamed `WalletsItem` to `WalletDtoWithTransaction`.
    - Renamed `CategoriesItem` to `CategoryDtoWithTransactions`.
    - Renamed `TransactionItem` to `TransactionsDtoWithWallet`.
    - Introduced `WalletDto`, `CategoryDto`, and `TransactionsDto` to replace various "Item" and legacy data classes.
- **Package Reorganization**:
    - Moved `GetTransactionsResponse` and related transaction DTOs to the `id.my.sendiko.fintrack.transaction.data.dto` package.
- **Clean up**:
    - Removed redundant `Category`, `Wallet`, and `TransactionsItem` classes across `transaction`, `category`, and `wallet` packages.
    - Deleted `PostWalletItem` in favor of the unified `WalletDto`.
- **Integration Updates**:
    - Updated `ApiService`, `KtorClient`, and `DashboardRepository` to use the new DTO names and locations.
    - Updated `LoginRepositoryImpl` and `RegisterRepositoryImpl` to map from `UserDto`.
    - Standardized API response models (`GetCategoriesResponse`, `PostWalletResponse`, `LoginResponse`, `RegisterResponse`, `GetWalletsResponse`) to use the new DTO structures.
