+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/auth/login/data/dto/LoginResponse.kt
@@ -1,6 +1,6 @@
 package id.my.sendiko.fintrack.auth.login.data.dto
 
-import id.my.sendiko.fintrack.auth.core.data.dto.UserItem
+import id.my.sendiko.fintrack.auth.core.data.dto.UserDto
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
@@ -11,7 +11,7 @@ data class LoginResponse(
 	val message: String,
 
 	@SerialName("user")
-	val userItem: UserItem,
+	val userDto: UserDto,
 
 	@SerialName("status")
 	val status: Int
