--- shared/src/commonMain/kotlin/id/my/sendiko/fintrack/category/data/CategoriesItem.kt
@@ -1,26 +1,0 @@
-package id.my.sendiko.fintrack.category.data
-
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
-
-@Serializable
-data class CategoriesItem(
-
-	@SerialName("createdAt")
-	val createdAt: String,
-
-	@SerialName("name")
-	val name: String,
-
-	@SerialName("id")
-	val id: String,
-
-	@SerialName("transactions")
-	val transactions: List<TransactionsItem>,
-
-	@SerialName("userId")
-	val userId: String,
-
-	@SerialName("updatedAt")
-	val updatedAt: String
-)
