--- shared/src/commonMain/kotlin/id/my/sendiko/fintrack/transaction/data/TransactionsItem.kt
@@ -1,44 +1,0 @@
-package id.my.sendiko.fintrack.transaction.data
-
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
-
-@Serializable
-data class TransactionsItem(
-
-	@SerialName("walletId")
-	val walletId: String,
-
-	@SerialName("createdAt")
-	val createdAt: String,
-
-	@SerialName("amount")
-	val amount: Int,
-
-	@SerialName("deletedAt")
-	val deletedAt: String?,
-
-	@SerialName("wallet")
-	val wallet: Wallet,
-
-	@SerialName("name")
-	val name: String,
-
-	@SerialName("id")
-	val id: String,
-
-	@SerialName("type")
-	val type: String,
-
-	@SerialName("category")
-	val category: Category,
-
-	@SerialName("userId")
-	val userId: String,
-
-	@SerialName("categoryId")
-	val categoryId: String,
-
-	@SerialName("updatedAt")
-	val updatedAt: String
-)
