+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/transaction/data/dto/TransactionsDtoWithWallet.kt
@@ -1,10 +1,12 @@
-package id.my.sendiko.fintrack.transaction.data
+package id.my.sendiko.fintrack.transaction.data.dto
 
+import id.my.sendiko.fintrack.category.data.CategoryDto
+import id.my.sendiko.fintrack.wallet.core.data.dto.WalletDto
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
 @Serializable
-data class TransactionItem(
+data class TransactionsDtoWithWallet(
 
 	@SerialName("walletId")
 	val walletId: String,
@@ -19,7 +21,7 @@ data class TransactionItem(
 	val deletedAt: String?,
 
 	@SerialName("wallet")
-	val wallet: Wallet,
+	val wallet: WalletDto,
 
 	@SerialName("name")
 	val name: String,
@@ -31,7 +33,7 @@ data class TransactionItem(
 	val type: String,
 
 	@SerialName("category")
-	val category: Category,
+	val category: CategoryDto,
 
 	@SerialName("userId")
 	val userId: String,
