+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/wallet/core/data/dto/getdetails/GetWalletsResponse.kt
@@ -1,18 +1,18 @@
 package id.my.sendiko.fintrack.wallet.core.data.dto.getdetails
 
-import id.my.sendiko.fintrack.wallet.core.data.dto.WalletsItem
+import id.my.sendiko.fintrack.wallet.core.data.dto.WalletDtoWithTransaction
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
 @Serializable
 data class GetWalletsResponse(
 
-	@SerialName("wallets")
-	val wallets: List<WalletsItem>,
+    @SerialName("wallets")
+	val wallets: List<WalletDtoWithTransaction>,
 
-	@SerialName("message")
+    @SerialName("message")
 	val message: String,
 
-	@SerialName("status")
+    @SerialName("status")
 	val status: Int
 )
