+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/wallet/core/data/dto/post/PostWalletResponse.kt
@@ -1,5 +1,6 @@
 package id.my.sendiko.fintrack.wallet.core.data.dto.post
 
+import id.my.sendiko.fintrack.wallet.core.data.dto.WalletDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -7,7 +8,7 @@ import kotlinx.serialization.Serializable
 data class PostWalletResponse(
 
     @SerialName("wallet")
-    val wallet: PostWalletItem,
+    val wallet: WalletDto,
 
     @SerialName("message")
     val message: String,
